completed my next task

+++ src/main/java/com/example/task/response/TransactionResponse.java
@@ -2,16 +2,17 @@ package com.example.task.response;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
 public class TransactionResponse {
 
     private LocalDateTime transactionDate;
-    private Long amount;
+    private BigDecimal amount;
     private String transactionDetails;
 
-    public TransactionResponse(LocalDateTime transactionDate, Long amount, String transactionDetails) {
+    public TransactionResponse(LocalDateTime transactionDate, BigDecimal amount, String transactionDetails) {
         this.transactionDate = transactionDate;
         this.amount = amount;
         this.transactionDetails = transactionDetails;
